Update hash_format function, return hash in S3 responses
